[GB.QT4] * NEW: Control.PopupMenu is a new property that allows to define the name   of the menu that will be automatically popped up when the user clicks on   the control with the right mouse button, or press the menu key.   Beware that if the Menu event handler is already defined for the control,   the popup menu will not open.
[GB.GTK]
* NEW: Control.PopupMenu is a new property that allows to define the name 
  of the menu that will be automatically popped up when the user clicks on
  the control with the right mouse button, or press the menu key.
  Beware that if the Menu event handler is already defined for the control, 
  the popup menu will not open.


git-svn-id: svn://localhost/gambas/trunk@3064 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.gtk/src/CWindow.cpp
@@ -37,8 +37,8 @@ bool CWINDOW_Embedded = false;
 
 CWINDOW *CWINDOW_Active = NULL;
 
-static CWINDOW *MAIN_Window=NULL;
-static long MODAL_windows=0;
+static CWINDOW *MAIN_Window = NULL;
+static int MODAL_windows = 0;
 
 DECLARE_EVENT(EVENT_Open);
 DECLARE_EVENT(EVENT_Show);
